+++ js/main.js
@@ -1,1 +1,1 @@
-import './thumbnails.js';
+import './gallery.js';
